Customer & Cart Unit tests (#107)
* refactor: encapsulate email setters into one function

* refactor: run code clean-up

* feat: add Unit|Mock test for CartService class.

* feat: define 2 custom exceptions NegativeQuantityException and NullCartException

* refactor: update CartService to throw exceptions

* feat: define more custom exceptions for Null[ cartId, cartItemId, cartItem]

* refactor: adapt CartService and CartServiceUnitTest to new custom exceptions created

* refactor: remove NullCartIdException.java and NullCartItemIdException.java and unify them into NullIdException

* feat: define new exceptions for Customer

* feat: add exceptions to CustomerService

* feat: add initial tests for CustomerServiceUnitTest

+++ src/main/java/com/vodafone/service/CartService.java
@@ -1,15 +1,12 @@
 package com.vodafone.service;
 
+import com.vodafone.exception.*;
 import com.vodafone.model.Cart;
 import com.vodafone.model.CartItem;
 import com.vodafone.model.Order;
 import com.vodafone.model.OrderItem;
 import com.vodafone.repository.cart.ICartRepository;
 import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import org.hibernate.HibernateException;
-import org.hibernate.Session;
-import org.hibernate.Transaction;
 import org.springframework.stereotype.Service;
 
 import java.sql.Date;
@@ -25,21 +22,34 @@ public class CartService {
 
 
     public boolean create(Cart entity) {
+        if (entity == null)
+            throw new NullCartException("Null cart provided");
         return cartRepository.create(entity);
     }
 
 
     public boolean update(Long id, Cart updatedEntity) {
+//        if (updatedEntity == null || get(id) == null)
+        if (id == null)
+            throw new NullIdException("Null cart id is provided");
+        if (updatedEntity == null)
+            throw new NullCartException("Null cart provided");
         return cartRepository.update(id, updatedEntity);
     }
 
 
     public boolean delete(Long id) {
+//        if (get(id) == null)
+//          throw new NullPointerException("No cart exists with this id");
+        if (id == null)
+            throw new NullIdException("Null cart id is provided");
         return cartRepository.delete(id);
     }
 
 
     public Cart get(Long id) {
+        if (id == null)
+            throw new NullIdException("Null cart id is provided");
         return cartRepository.get(id);
     }
 
@@ -50,6 +60,15 @@ public class CartService {
 
 
     public boolean removeItem(Long cartId, Long itemId) {
+//        if (get(cartId) == null)
+//            throw new NullPointerException("No cart exists with this id");
+        if (cartId == null)
+            throw new NullIdException("Null cart id is provided");
+        if (itemId == null)
+            throw new NullIdException("Null item id is provided");
+//        long matches = get(cartId).getItems().stream().filter(item -> item.getId().equals(itemId)).count();
+//        if (matches == 0)
+//            throw new NulLCartItemException("No item exists with this id");
         return cartRepository.removeItem(cartId, itemId);
     }
 
@@ -69,7 +88,7 @@ public class CartService {
         //iterate over each cart item to transform it to order item.
         float total = 0f;
         for (CartItem cartItem : cart.getItems()) {
-            if(cartItem.getQuantity()>0) {
+            if (cartItem.getQuantity() > 0) {
                 OrderItem orderItem = new OrderItem();
                 orderItem.setOrder(order);
                 //check available quantity in stock
@@ -96,26 +115,52 @@ public class CartService {
     }
 
     public boolean clearCart(Long cartId) {
+//        if (get(cartId) == null)
+//            throw new NullPointerException("Null cart id is provided");
+        if (cartId == null)
+            throw new NullIdException("Null cart id is provided");
         return cartRepository.clearCart(cartId);
     }
 
-    public int addItem(Long cartId, CartItem item) {
+    public boolean addItem(Long cartId, CartItem item) {
+//        if (cartId == null)
+//            throw new NullPointerException("Null cart id is provided");
+        if (cartId == null)
+            throw new NullIdException("Null cart id is provided");
+        if (item == null)
+            throw new NullCartItemException("Null cart item is provided");
         return cartRepository.addItem(cartId, item);
     }
 
     public List<CartItem> getCartItems(Long cartId) {
+        if (cartId == null)
+            throw new NullIdException("Null cart id is provided");
         return cartRepository.getCartItems(cartId);
     }
 
     public int setProductQuantity(Long cartId, Long itemId, int newQuantity) {
+        if (cartId == null)
+            throw new NullIdException("Null cart id is provided");
+        if (itemId == null)
+            throw new NullIdException("Null cart item is provided");
+        if (newQuantity < 0)
+            throw new NegativeQuantityException("Negative quantity provided");
         return cartRepository.setProductQuantity(cartId, itemId, newQuantity);
     }
 
-    public int incrementProductQuantity(Long cartId, Long productId, int quantity) {
-        return cartRepository.incrementProductQuantity(cartId, productId, quantity);
+    public int incrementProductQuantity(Long cartId, Long itemId, int quantity) {
+        if (cartId == null)
+            throw new NullIdException("Null cart id is provided");
+        if (itemId == null)
+            throw new NullIdException("Null cart item is provided");
+        return cartRepository.incrementProductQuantity(cartId, itemId, quantity);
     }
 
-    public int decrementProductQuantity(Long cartId, Long productId) {
-        return cartRepository.decrementProductQuantity(cartId, productId);
+    public int decrementProductQuantity(Long cartId, Long itemId) {
+        if (cartId == null)
+            throw new NullIdException("Null cart id is provided");
+        if (itemId == null)
+            throw new NullIdException("Null cart item is provided");
+        return cartRepository.decrementProductQuantity(cartId, itemId);
     }
 }
